reactor: イベント一覧を別コンポーネントに切り出し
